requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain-core>=0.2.22,<0.3
langgraph==0.1.13
langchain-openai>=0.0.5
langchain-community>=0.0.9
openai>=1.3.0
azure-storage-blob==12.19.0
azure-search-documents==11.4.0
pyodbc==5.0.1
sqlalchemy==2.0.23
pydantic==2.5.0
redis==5.0.1
tenacity==8.2.3
structlog==23.2.0
python-multipart==0.0.6
asyncio-mqtt==0.16.1
pytest==7.4.3
pytest-asyncio==0.21.1